requirements.txt
python-docx>=1.1.0
google-genai>=0.1.0
python-dotenv>=1.0.0